+++ src/GitHub_Updater/Branch.php
@@ -79,7 +79,7 @@ class Branch {
 	public function set_branch_on_switch( $repo ) {
 		$this->cache = Singleton::get_instance( 'API_PseudoTrait' )->get_repo_cache( $repo );
 
-		if ( isset( $_GET['action'], $this->cache['branches'] ) &&
+		if ( isset( $_GET['action'], $_GET['rollback'], $this->cache['branches'] ) &&
 		     ( 'upgrade-plugin' === $_GET['action'] || 'upgrade-theme' === $_GET['action'] )
 		) {
 			$current_branch = array_key_exists( $_GET['rollback'], $this->cache['branches'] )