+++ src/GitHub_Updater/GitHub_API.php
@@ -474,6 +474,12 @@ class GitHub_API extends API implements API_Interface {
 				$response->message = 'No release asset found';
 			}
 
+			if ( is_wp_error( $response ) ) {
+				Singleton::get_instance( 'Messages' )->create_error_message( $response );
+
+				return false;
+			}
+
 			if ( $response ) {
 				add_filter( 'http_request_args', array( &$this, 'set_github_release_asset_header' ) );
 
@@ -685,7 +691,9 @@ class GitHub_API extends API implements API_Interface {
 				: static::$options['github_enterprise_token'];
 		}
 
-		$install['download_link'] = add_query_arg( 'access_token', $token, $install['download_link'] );
+		if ( ! empty( $token ) ) {
+			$install['download_link'] = add_query_arg( 'access_token', $token, $install['download_link'] );
+		}
 
 		if ( ! empty( static::$options['github_access_token'] ) ) {
 			unset( $install['options']['github_access_token'] );