+++ src/GitHub_Updater/Plugin.php
@@ -155,9 +155,9 @@ class Plugin extends Base {
 				$git_plugin['name']                    = $plugin_data['Name'];
 				$git_plugin['local_version']           = strtolower( $plugin_data['Version'] );
 				$git_plugin['sections']['description'] = $plugin_data['Description'];
-				$git_plugin['languages']               = ! empty( $header['languages'] ) ? $header['languages'] : null;
-				$git_plugin['ci_job']                  = ! empty( $header['ci_job'] ) ? $header['ci_job'] : null;
-				$git_plugin['release_asset']           = 'true' === $plugin_data['Release Asset'];
+				$git_plugin['languages']               = $header['languages'];
+				$git_plugin['ci_job']                  = $header['ci_job'];
+				$git_plugin['release_asset']           = $header['release_asset'];
 				$git_plugin['broken']                  = ( empty( $header['owner'] ) || empty( $header['repo'] ) );
 
 				$git_plugin['banners']['high'] =