+++ src/GitHub_Updater/Theme.php
@@ -161,9 +161,9 @@ class Theme extends Base {
 				$git_theme['sections']['description'] = $theme->get( 'Description' );
 				$git_theme['local_path']              = get_theme_root() . '/' . $git_theme['repo'] . '/';
 				$git_theme['branch']                  = $branch ?: 'master';
-				$git_theme['languages']               = ! empty( $header['languages'] ) ? $header['languages'] : null;
-				$git_theme['ci_job']                  = ! empty( $header['ci_job'] ) ? $header['ci_job'] : null;
-				$git_theme['release_asset']           = 'true' === $theme->get( 'Release Asset' );
+				$git_theme['languages']               = $header['languages'];
+				$git_theme['ci_job']                  = $header['ci_job'];
+				$git_theme['release_asset']           = $header['release_asset'];
 				$git_theme['broken']                  = ( empty( $header['owner'] ) || empty( $header['repo'] ) );
 
 				break;