+++ CHANGES.md
@@ -1,5 +1,10 @@
 #### [unreleased]
 
+#### 7.4.3 / 2017-11-07
+* set all extra header values in `Base::parse_extra_headers()`
+* added more error messaging for `class WP_Error`
+* fixed some issues with GitHub Release Assets
+
 #### 7.4.2 / 2017-10-25
 * added check to see if wp-cron is updating and if not send and error message
 * fix for WP-CLI updating for private Bitbucket repos, thanks @v8-ict